Reworking sys info implementation
Making sure that each device type correctly constructs it's own unique
system information.

Also removed the concept of a generic "device child" and "device child
action" given that currently I am only aware of the HS300 device having
children, and their system information may be specific to that device
type.
