+++ tplinkcloud/hs100.py
@@ -38,3 +38,7 @@ class HS100(TPLinkEMeterDevice):
     def __init__(self, client, device_id, device_info):
         super().__init__(client, device_id, device_info)
         self.model_type = TPLinkDeviceType.HS100
+
+    def get_sys_info(self):
+        sys_info = self._get_sys_info()
+        return HS100SysInfo(sys_info)
