+++ tplinkcloud/hs300.py
@@ -1,7 +1,6 @@
 from .device_type import TPLinkDeviceType
 from .emeter_device import TPLinkEMeterDevice
-from .hs300_child import HS300Child
-from .device_child_info import TPLinkDeviceChildInfo
+from .hs300_child import HS300Child, HS300ChildSysInfo
 
 class HS300SysInfo:
 
@@ -22,7 +21,7 @@ class HS300SysInfo:
         self.mac = sys_info.get('mac')
         self.updating = sys_info.get('updating')
         self.led_off = sys_info.get('led_off')
-        self.children = sys_info.get('children')
+        self.children = [HS300ChildSysInfo(child_info) for child_info in sys_info.get('children')]
         self.child_num = sys_info.get('child_num')
         self.err_code = sys_info.get('err_code')
 
@@ -33,14 +32,17 @@ class HS300(TPLinkEMeterDevice):
         self.model_type = TPLinkDeviceType.HS300
     
     def get_children(self):
-        sys_info = HS300SysInfo(self.get_sys_info())
+        sys_info = self.get_sys_info()
         children = []
         for child_info in sys_info.children:
-            device_child_info = TPLinkDeviceChildInfo(child_info)
-            device_child = HS300Child(self.client, self.device_id, device_child_info.id, device_child_info)
+            device_child = HS300Child(self.client, sys_info.device_id, child_info.id, child_info)
             children.append(device_child)
         return children
 
     # An override of an identified TPLinkDevice
     def has_children(self):
         return True
+
+    def get_sys_info(self):
+        sys_info = self._get_sys_info()
+        return HS300SysInfo(sys_info)
